pyproject.toml
[build-system]
requires = ["setuptools", "setuptools-scm"]
build-backend = "setuptools.build_meta"

[project]
name = "dupicolib"
version = "0.2.0"
description = "Library to interface with the dupico hardware"
authors = [
   { name = "Fabio Battaglia", email = "hkzlabnet@gmail.com" }
]
keywords = ["dupico", "DuPAL", "dumping"]
readme = "README.md"
license = { text = "CC BY-SA 4.0 DEED" }
classifiers = [
    "Programming Language :: Python :: 3"
]
requires-python = ">=3.12"
dependencies = [
    "pyserial ~= 3.5",
]

[tool.setuptools]
packages = [ "dupicolib" ]
py-modules = [ "__init__" ]

[project.urls]
repository = "https://github.com/DuPAL-PAL-DUmper/dupicolib"